+++ lib/home/home.dart
@@ -3,7 +3,7 @@ import "package:app/home/spends.dart";
 import "package:app/utility/schema/methods.dart";
 import "package:flutter/material.dart";
 import "package:app/cards/balance.dart";
-import "package:app/cards/addSpend.dart";
+import "package:app/ops/addSpend.dart";
 import "package:font_awesome_flutter/font_awesome_flutter.dart";
 
 class Home extends StatefulWidget {
