+++ lib/onboarding/welcome.dart
@@ -1,6 +1,6 @@
-import "package:app/cards/addSpend.dart";
-import "package:app/cards/addIncome.dart";
-import "package:app/cards/addCategories.dart";
+import "package:app/ops/addSpend.dart";
+import "package:app/ops/addIncome.dart";
+import "package:app/ops/addCategories.dart";
 import "package:app/utility/schema/methods.dart";
 import "package:app/navigation.dart";
 import "package:flutter/material.dart";
