+++ lib/ops/addCategories.dart
