+++ lib/ops/addSubscription.dart
